+++ blender-qmm/blender-qmm/mats/SteelRoughness.py
@@ -23,39 +23,27 @@ class SteelRoughnessGroup(bpy.types.Operator):
 
         # create outputs
         if bpy.app.version < (4, 0, 0):
-            sr_group.outputs.new('NodeSocketFloat', 'Carbon Steel')                    #0
-            sr_group.outputs.new('NodeSocketFloat', 'Stainless Steel')                 #1
-            sr_group.outputs.new('NodeSocketFloat', 'Alloy Steel')                     #2
-            sr_group.outputs.new('NodeSocketFloat', 'Tool Steel')                      #3
-            sr_group.outputs.new('NodeSocketFloat', 'Spring Steel')                    #4
-            sr_group.outputs.new('NodeSocketFloat', 'Structural Steel')                #5
-            sr_group.outputs.new('NodeSocketFloat', 'High-Strength, Low-Alloy Steel')  #6
-            sr_group.outputs.new('NodeSocketFloat', 'Maraging Steel')                  #7
-            sr_group.outputs.new('NodeSocketFloat', 'Weathering Steel')                #8
-            sr_group.outputs.new('NodeSocketFloat', 'Free-machining Steel')            #9
+            sr_group.outputs.new('NodeSocketFloat', 'Maraging, Weathering Steel')      #0
+            sr_group.outputs.new('NodeSocketFloat', 'Free-Machining Steel')            #1
+            sr_group.outputs.new('NodeSocketFloat', 'Spring, Stainless, HSLA Steel')   #2
+            sr_group.outputs.new('NodeSocketFloat', 'Structural, Tool Steel')          #3
+            sr_group.outputs.new('NodeSocketFloat', 'Alloy Steel')                     #4
+            sr_group.outputs.new('NodeSocketFloat', 'Carbon Steel')                    #5
         else:
-            sr_group.interface.new_socket(name="Carbon Steel", in_out='OUTPUT', socket_type='NodeSocketFloat')
-            sr_group.interface.new_socket(name="Stainless Steel", in_out='OUTPUT', socket_type='NodeSocketFloat')
+            sr_group.interface.new_socket(name="Maraging, Weathering Steel", in_out='OUTPUT', socket_type='NodeSocketFloat')
+            sr_group.interface.new_socket(name="Free-Machining Steel", in_out='OUTPUT', socket_type='NodeSocketFloat')
+            sr_group.interface.new_socket(name="Spring, Stainless, HSLA Steel", in_out='OUTPUT', socket_type='NodeSocketFloat')
+            sr_group.interface.new_socket(name="Structural, Tool Steel", in_out='OUTPUT', socket_type='NodeSocketFloat')
             sr_group.interface.new_socket(name="Alloy Steel", in_out='OUTPUT', socket_type='NodeSocketFloat')
-            sr_group.interface.new_socket(name="Tool Steel", in_out='OUTPUT', socket_type='NodeSocketFloat')
-            sr_group.interface.new_socket(name="Spring Steel", in_out='OUTPUT', socket_type='NodeSocketFloat')
-            sr_group.interface.new_socket(name="Structural Steel", in_out='OUTPUT', socket_type='NodeSocketFloat')
-            sr_group.interface.new_socket(name="High-Strength, Low-Alloy Steel", in_out='OUTPUT', socket_type='NodeSocketFloat')
-            sr_group.interface.new_socket(name="Maraging Steel", in_out='OUTPUT', socket_type='NodeSocketFloat')
-            sr_group.interface.new_socket(name="Weathering Steel", in_out='OUTPUT', socket_type='NodeSocketFloat')
-            sr_group.interface.new_socket(name="Free-machining Steel", in_out='OUTPUT', socket_type='NodeSocketFloat')
+            sr_group.interface.new_socket(name="Carbon Steel", in_out='OUTPUT', socket_type='NodeSocketFloat')
 
         #set values
-        group_out.inputs[0].default_value = 0.5
-        group_out.inputs[1].default_value = 0.2
-        group_out.inputs[2].default_value = 0.4
+        group_out.inputs[0].default_value = 0.05
+        group_out.inputs[1].default_value = 0.1
+        group_out.inputs[2].default_value = 0.2
         group_out.inputs[3].default_value = 0.3
-        group_out.inputs[4].default_value = 0.2
-        group_out.inputs[5].default_value = 0.3
-        group_out.inputs[6].default_value = 0.2
-        group_out.inputs[7].default_value = 0.05
-        group_out.inputs[8].default_value = 0.05
-        group_out.inputs[9].default_value = 0.1
+        group_out.inputs[4].default_value = 0.4
+        group_out.inputs[5].default_value = 0.5
 
     def make_node(self, group, arg1, arg2, arg3):
         result = group.nodes.new(arg1)