+++ blender-qmm/blender-qmm/mats/extras/AsphaltBleached.py
@@ -74,14 +74,22 @@ class QMMAsphaltBleached(bpy.types.Operator):
         BSDF = nodes.get('Principled BSDF')
         BSDF.distribution = 'MULTI_GGX'
         BSDF.location = (-700, 300)
-        BSDF.inputs[9].default_value = 0.56
+        if bv < (4, 0, 0):
+            BSDF.inputs[9].default_value = 0.56
+        else:
+            BSDF.inputs[2].default_value = 0.56
+            BSDF.inputs[3].default_value = 1.635
         # BSDF.select = True
 
         # principledbsdf - cracks
         BSDF2 = make_node(nodes, 'ShaderNodeBsdfPrincipled', -700, -400)
         BSDF2.distribution = 'MULTI_GGX'
         BSDF2.inputs[0].default_value = (0.025, 0.01875, 0.01875, 1)
-        BSDF2.inputs[9].default_value = 1.0
+        if bv < (4, 0, 0):
+            BSDF2.inputs[9].default_value = 1.0
+        else:
+            BSDF2.inputs[2].default_value = 1.0
+            BSDF2.inputs[3].default_value = 1.52
 
         # maprange
         m_maprange = make_node(nodes, 'ShaderNodeMapRange', -1100, 100)
@@ -169,12 +177,14 @@ class QMMAsphaltBleached(bpy.types.Operator):
         m_val2.label = "Gravel Scale"
         m_val2.outputs[0].default_value = 1.0
 
-        # EnergyConservationGroup
+        # EnergyConservationGroups
         bpy.ops.node.ec_group_operator()
-        ec_group = set_ec(nodes, -900, 200, (0.2, 0.2, 0.2, 1), 0.56, 1.635, (0.01, 0.01, 0.01, 1))
-
-        # EnergyConservationGroup2
-        ec_group2 = set_ec(nodes, -900, -600, (0.025000, 0.018750, 0.018750, 1), 1, 1.52, (0.01, 0.01, 0.01, 1))
+        if bv < (4, 0, 0):
+            ec_group = set_ec(nodes, -900, 200, (0.2, 0.2, 0.2, 1), 0.56, 1.635, (0.01, 0.01, 0.01, 1))
+            ec_group2 = set_ec(nodes, -900, -600, (0.025000, 0.018750, 0.018750, 1), 1, 1.52, (0.01, 0.01, 0.01, 1))
+        else:
+            ec_group = set_ec(nodes, 0, 400, (0.2, 0.2, 0.2, 1), 0.56, 1.635, (0.01, 0.01, 0.01, 1))
+            ec_group2 = set_ec(nodes, 0, -200, (0.025000, 0.018750, 0.018750, 1), 1, 1.52, (0.01, 0.01, 0.01, 1))
 
         links = m_asphalt_b.node_tree.links.new
 
@@ -182,10 +192,11 @@ class QMMAsphaltBleached(bpy.types.Operator):
         links(m_maprange2.outputs[0], m_mixshader.inputs[0])
         links(BSDF.outputs[0], m_mixshader.inputs[1])
         links(BSDF2.outputs[0], m_mixshader.inputs[2])
-        links(m_maprange.outputs[0], ec_group.inputs[0])
         if bpy.app.version < (4, 0, 0):
+            links(m_maprange.outputs[0], ec_group.inputs[0])
             links(m_bump.outputs[0], BSDF.inputs[22])
         else:
+            links(m_maprange.outputs[0], BSDF.inputs[0])
             links(m_bump.outputs[0], BSDF.inputs[5])
         links(m_bump2.outputs[0], m_bump.inputs[3])
         links(m_maprange2.outputs[0], m_bump2.inputs[2])
@@ -218,25 +229,25 @@ class QMMAsphaltBleached(bpy.types.Operator):
             links(m_noise3.outputs[1], m_mix2.inputs[6])
             links(m_mapping.outputs[0], m_mix2.inputs[7])
 
-        links(ec_group.outputs[0], BSDF.inputs[0])
         if bpy.app.version < (4, 0, 0):
+            links(ec_group.outputs[0], BSDF.inputs[0])
             links(ec_group.outputs[1], BSDF.inputs[7])
             links(ec_group.outputs[2], BSDF.inputs[9])
             links(ec_group.outputs[4], BSDF.inputs[16])
-        else:
-            links(ec_group.outputs[1], BSDF.inputs[12])
-            links(ec_group.outputs[2], BSDF.inputs[2])
-            links(ec_group.outputs[4], BSDF.inputs[3])
+        # else:
+        #     links(ec_group.outputs[1], BSDF.inputs[12])
+        #     links(ec_group.outputs[2], BSDF.inputs[2])
+        #     links(ec_group.outputs[4], BSDF.inputs[3])
 
-        links(ec_group2.outputs[0], BSDF2.inputs[0])
         if bpy.app.version < (4, 0, 0):
+            links(ec_group2.outputs[0], BSDF2.inputs[0])
             links(ec_group2.outputs[1], BSDF2.inputs[7])
             links(ec_group2.outputs[2], BSDF2.inputs[9])
             links(ec_group2.outputs[4], BSDF2.inputs[16])
-        else:
-            links(ec_group2.outputs[1], BSDF2.inputs[12])
-            links(ec_group2.outputs[2], BSDF2.inputs[2])
-            links(ec_group2.outputs[4], BSDF2.inputs[3])
+        # else:
+        #     links(ec_group2.outputs[1], BSDF2.inputs[12])
+        #     links(ec_group2.outputs[2], BSDF2.inputs[2])
+        #     links(ec_group2.outputs[4], BSDF2.inputs[3])
 
         links(m_maprange2.outputs[0], m_lessthan.inputs[0])
         links(m_lessthan.outputs[0], m_disp.inputs[0])